server/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fsrs==3.1.0
idna==3.10
libretranslatepy==2.1.1
lxml==5.3.0
numpy==2.0.1
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
translate==3.6.1
tzdata==2024.1
urllib3==2.2.3
